Begin SRU
Interface and types for HoldingsLookupAdaptor in own source

The new SRU client implementation uses httpclient which is
a verbatim of that in illmock. Should httpclient be its own
module?

DRY: one App initialization in Init. utils.go only with
changes for running it in separate Go routine.

The App startup may returns error.