CROSSLINK-151 broadcast solution
at this point tied to event confirm-requester-msg. The list
of events can executed in any instance, but the final
comfirm-request-msg is sent to all eventbus instances.

A more elegant solution would be to use a more generic
broadcast flag or similar for the event rather than tying
to event name.