Always confirm suspended requester message
Check the last requester action on the transaction rather than on selected supplier
This ensures that even upon failure when sending a message to supplier,
the  pending requester message is confirmed.

There is a small change in logic because of this: selecting next supplier when messaging
the old one fails will now only resume if the last requester action is Request.
E.g if the requester sends a notification to the old supplier and that message fails to deliver
the request will be stuck with the supplier until the supplier takes action.

We may need a special message to select a new supplier.